+++ src/color/index.ts
@@ -17,6 +17,7 @@
  * limitations under the License.
  */
 
+export { catppuccin2ThemeColor } from "./catppuccin";
 export { display2GitHubColor, type DisplayColor } from "./display";
 export { gitea2ThemeVars, type GiteaColor } from "./gitea";
 export { github2ThemeColor, type GitHubColor } from "./github";