+++ src/index.ts
@@ -19,7 +19,7 @@
 
 import selectors from "./selectors";
 
-export { display2GitHubColor, github2ThemeColor } from "./color";
+export { display2GitHubColor, github2ThemeColor, catppuccin2ThemeColor } from "./color";
 export type { DisplayColor, GitHubColor, GitHubSyntax, GiteaColor, ThemeColor } from "./color";
 export { createTheme, defineTheme } from "./core";
 export { extractVarName, scaleColorLight } from "./functions";